combined nn and linear interpolation test into single file
